requirements.txt
altgraph==0.17.2
future==0.18.2
pefile==2021.9.3
Pillow==9.0.0
pyinstaller==4.8
pyinstaller-hooks-contrib==2022.0
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
PyQt5-stubs==5.15.2.0
pywin32-ctypes==0.2.0
